requirements.txt
fastapi[all]
uvicorn
google-cloud-firestore
google-auth
pydantic